refactor(app-paths): split the crate root into focused path modules
