+++ src-tauri/crates/app-paths/src/data_root.rs
@@ -1,0 +1,378 @@
+//! The `~/.orgii/` data root and the files and directories that live directly
+//! under it: databases, caches, settings, the OS Agent ("personal") family,
+//! sync stores, policy/MCP config, and the file-based session registry.
+
+use std::path::{Path, PathBuf};
+
+use crate::home::home_dir;
+
+/// Application data root: `~/.orgii/`.
+///
+/// Test override: setting `ORGII_HOME` redirects every path under the data
+/// root to that directory. Production code must never set it.
+///
+/// Fallback: `$TMPDIR/.orgii/` (only reached if `dirs::home_dir()` fails).
+pub fn orgii_root() -> PathBuf {
+    if let Ok(override_path) = std::env::var("ORGII_HOME") {
+        return PathBuf::from(override_path);
+    }
+    home_dir().join(".orgii")
+}
+
+/// Provider keys file: `~/.orgii/credentials.json`.
+///
+/// Filename retained from the legacy Python implementation; the user-facing
+/// terminology in code and UI is "key", not "credential".
+pub fn keys() -> PathBuf {
+    orgii_root().join("credentials.json")
+}
+
+/// Local MITM proxy directory: `~/.orgii/proxy/`.
+///
+/// Holds the root CA certificate (`ca.pem` + `ca-key.pem`) generated and
+/// consumed by `integrations::proxy::certificate_authority`. Callers create
+/// the directory on demand.
+pub fn proxy_dir() -> PathBuf {
+    orgii_root().join("proxy")
+}
+
+/// Main SQLite database file: `~/.orgii/sessions.db`.
+///
+/// Hosts session events, CLI agent state, inbox, dev records, lineage,
+/// orchestrator state, plan approvals, and the agent-core unified session
+/// layer. The `database` workspace crate resolves this path on every
+/// `get_connection()` call, so leaf crates that need to open the file (e.g.
+/// the test harness) can call this helper directly without re-deriving the
+/// join.
+pub fn sessions_db() -> PathBuf {
+    orgii_root().join("sessions.db")
+}
+
+/// Durable append-only shell replay artifacts: `~/.orgii/shell-replays/`.
+///
+/// Kept under the same `ORGII_HOME`-aware root as `sessions.db` so desktop,
+/// headless/API execution, range reads, tests, and session deletion always
+/// resolve the identical lifecycle-owned directory.
+pub fn shell_replays_dir() -> PathBuf {
+    orgii_root().join("shell-replays")
+}
+
+/// Privacy-filtered session-provenance hook inbox:
+/// `~/.orgii/session-provenance/inbox/`.
+///
+/// External agent hooks write small, versioned envelopes here instead of
+/// opening `sessions.db` directly. The desktop process drains the inbox and
+/// owns all canonical SQLite writes.
+pub fn session_provenance_inbox_dir() -> PathBuf {
+    orgii_root().join("session-provenance").join("inbox")
+}
+
+/// Live agent-status loopback endpoint descriptor:
+/// `~/.orgii/session-provenance/status-endpoint.json`.
+///
+/// Written atomically by the desktop each launch (`{version, port, token,
+/// pid, startedAt}`), re-read by every hook subprocess invocation so CLI
+/// sessions that outlive an Orgii restart reach the new server/token. Never
+/// deleted on shutdown — a dead server just refuses the TCP connect.
+pub fn agent_status_endpoint_path() -> PathBuf {
+    orgii_root()
+        .join("session-provenance")
+        .join("status-endpoint.json")
+}
+
+/// Live agent-status last-status cache:
+/// `~/.orgii/session-provenance/last-status.json`.
+///
+/// Debounced snapshot of the in-memory live-status map, hydrated on startup
+/// (TTL-filtered) for UI continuity across restarts. Owner-only permissions;
+/// never mirrored into `sessions.db`.
+pub fn agent_status_cache_path() -> PathBuf {
+    orgii_root()
+        .join("session-provenance")
+        .join("last-status.json")
+}
+
+/// Project & work-item SQLite database: `~/.orgii/projects/projects.db`.
+///
+/// Kept separate from `sessions_db` so cross-device sync and manual
+/// export/import treat the project store as a self-contained file. The
+/// parent directory must be created on demand by callers (the path itself
+/// is just a `join`).
+pub fn projects_db() -> PathBuf {
+    orgii_root().join("projects").join("projects.db")
+}
+
+/// User settings JSONC file: `~/.orgii/settings.jsonc`.
+///
+/// VS-Code-style: a single human-editable JSONC file the user (or an agent)
+/// can edit directly. The `settings` workspace crate watches this path for
+/// external modifications.
+pub fn settings() -> PathBuf {
+    orgii_root().join("settings.jsonc")
+}
+
+/// Settings JSON-Schema file: `~/.orgii/settings-schema.json`.
+///
+/// Generated by the `settings` crate; consumed by the IDE and the frontend's
+/// settings editor for autocomplete / validation.
+pub fn settings_schema() -> PathBuf {
+    orgii_root().join("settings-schema.json")
+}
+
+/// Screenshot capture directory: `~/.orgii/screenshots/`.
+///
+/// `shared_state::ScreenshotStore` writes captures here; both `agent_core`
+/// and `browser` enumerate the directory at runtime. Callers create the dir
+/// on demand.
+pub fn screenshots_dir() -> PathBuf {
+    orgii_root().join("screenshots")
+}
+
+/// LSP server discovery cache: `~/.orgii/lsp_cache.json`.
+///
+/// Read by the `lsp` crate's command layer, written by its install pipeline.
+pub fn lsp_cache() -> PathBuf {
+    orgii_root().join("lsp_cache.json")
+}
+
+/// Lint tool discovery cache: `~/.orgii/lint_cache.json`.
+pub fn lint_cache() -> PathBuf {
+    orgii_root().join("lint_cache.json")
+}
+
+/// Provider protocol capability cache: `~/.orgii/provider-capabilities.json`.
+pub fn provider_capabilities_cache() -> PathBuf {
+    orgii_root().join("provider-capabilities.json")
+}
+
+/// Usage diagnostics directory: `~/.orgii/diagnostics/`.
+pub fn diagnostics_dir() -> PathBuf {
+    orgii_root().join("diagnostics")
+}
+
+/// LSP server binaries installed by auto-install: `~/.orgii/lsp-bin/`.
+pub fn lsp_bin_dir() -> PathBuf {
+    orgii_root().join("lsp-bin")
+}
+
+/// Sidecar binaries downloaded at runtime: `~/.orgii/bin/`.
+///
+/// Peekaboo, agent-browser, and bundled git live here after the first-run
+/// download so they are never bundled in the notarized `.app` bundle.
+pub fn sidecar_bin_dir() -> PathBuf {
+    orgii_root().join("bin")
+}
+
+/// Global LSP configuration: `~/.orgii/lsp.json`.
+pub fn lsp_config() -> PathBuf {
+    orgii_root().join("lsp.json")
+}
+
+/// System-dependency probe cache: `~/.orgii/dependencies.json`.
+///
+/// Written by `infrastructure::platform::dependencies::detect_system_dependencies`.
+/// Read by leaf crates (e.g. `lsp`) that need to gate features on optional
+/// runtime tools without re-scanning. Exposing the path here lets those
+/// crates read the cache file without depending on the platform module.
+pub fn dependencies_cache() -> PathBuf {
+    orgii_root().join("dependencies.json")
+}
+
+/// Delete the dependency probe cache so the next call to
+/// `detect_system_dependencies` rescans from scratch.
+///
+/// Call this once at startup after `augment_path_from_shell()` so that a
+/// stale cache written under the old (stripped) PATH is not served back to
+/// the frontend as if it were still valid.
+pub fn clear_dependencies_cache() {
+    let path = dependencies_cache();
+    if path.exists() {
+        if let Err(err) = std::fs::remove_file(&path) {
+            tracing::warn!(
+                path = %path.display(),
+                error = %err,
+                "[app_paths] could not clear dependency cache; stale probe results may persist"
+            );
+        }
+    }
+}
+
+/// Per-repo Merkle snapshots root: `~/.orgii/merkle/`.
+///
+/// Read/written by the `search` crate's merkle workspace cache. Hoisted
+/// here so the crate doesn't depend back into the `app` crate.
+pub fn merkle_root() -> PathBuf {
+    orgii_root().join("merkle")
+}
+
+/// Local embedding/model downloads: `~/.orgii/models/`.
+pub fn models_dir() -> PathBuf {
+    orgii_root().join("models")
+}
+
+/// Semantic-search USearch index: `~/.orgii/semantic_index/`.
+pub fn semantic_index_dir() -> PathBuf {
+    orgii_root().join("semantic_index")
+}
+
+/// Plugin package root: `~/.orgii/extensions/`.
+///
+/// The directory name is retained on disk because plugin packages were stored
+/// there before the VS Code extension host was archived.
+pub fn extensions_dir() -> PathBuf {
+    orgii_root().join("extensions")
+}
+
+/// Plugin secret storage: `~/.orgii/extension_secrets.json`.
+pub fn extension_secrets() -> PathBuf {
+    orgii_root().join("extension_secrets.json")
+}
+
+// ============================================
+// OS Agent ("personal") path family
+// ============================================
+//
+// Hoisted here so `crates/agent-core` can resolve them without depending
+// back into the `app` crate. Originally extracted as PR 1 of the
+// agent_core extraction.
+
+/// OS Agent ("personal") root: `~/.orgii/personal/`.
+///
+/// Single source of truth for the OS Agent. Everything that belongs to
+/// the OS Agent — its automations, rules, skills, workspace — lives
+/// under this directory. SDE / workspace-scoped sessions never read or write
+/// here.
+pub fn personal_root() -> PathBuf {
+    orgii_root().join("personal")
+}
+
+/// OS Agent default workspace: `~/.orgii/personal/workspace/`.
+///
+/// Working directory where the OS Agent reads/writes files. Channel
+/// sessions (telegram/feishu/wecom/...) default to this directory when
+/// no workspace is attached.
+pub fn personal_workspace() -> PathBuf {
+    personal_root().join("workspace")
+}
+
+/// OS Agent automation rules: `~/.orgii/personal/automations.json`.
+pub fn personal_automations() -> PathBuf {
+    personal_root().join("automations.json")
+}
+
+/// OS Agent rules directory: `~/.orgii/personal/rules/`.
+///
+/// Markdown policy files that apply only to the OS Agent — never loaded
+/// for workspace-scoped SDE sessions.
+pub fn personal_rules_dir() -> PathBuf {
+    personal_root().join("rules")
+}
+
+/// OS Agent rules config: `~/.orgii/personal/rules-config.json`.
+pub fn personal_rules_config() -> PathBuf {
+    personal_root().join("rules-config.json")
+}
+
+/// Custom agent definitions (global, not OS-Agent-scoped):
+/// `~/.orgii/agent-definitions.json`.
+///
+/// Used by SDE, OS Agent, and any custom agent — intentionally outside
+/// `personal/` because it is not OS-Agent-specific.
+pub fn agent_definitions() -> PathBuf {
+    orgii_root().join("agent-definitions.json")
+}
+
+/// Agent organizations (global): `~/.orgii/agent-orgs.json`.
+pub fn agent_orgs() -> PathBuf {
+    orgii_root().join("agent-orgs.json")
+}
+
+/// Builtin-agent overrides overlay: `~/.orgii/builtin-overrides.json`.
+///
+/// User-writable overlay for `builtin:*` agent definitions. Loaded on
+/// top of the compiled-in builtin at runtime so users can re-pin
+/// models, tweak tool lists, and adjust policies without forking the
+/// Rust source. Storage contract §12 "Conflict C".
+pub fn builtin_overrides() -> PathBuf {
+    orgii_root().join("builtin-overrides.json")
+}
+
+/// Integrations config (Telegram/Discord/Feishu/...): `~/.orgii/integrations.json`.
+pub fn integrations() -> PathBuf {
+    orgii_root().join("integrations.json")
+}
+
+/// Root of the per-session file history archive: `~/.orgii/file-history/`.
+pub fn file_history_root() -> PathBuf {
+    orgii_root().join("file-history")
+}
+
+/// Per-session file history dir: `~/.orgii/file-history/<session_id>/`.
+pub fn file_history_dir(session_id: &str) -> PathBuf {
+    file_history_root().join(session_id)
+}
+
+/// Global project sync connection account store: `~/.orgii/sync_connections.json`.
+pub fn sync_connections() -> PathBuf {
+    orgii_root().join("sync_connections.json")
+}
+
+/// Global project sync connection token store: `~/.orgii/sync_connection_tokens.json`.
+pub fn sync_connection_tokens() -> PathBuf {
+    orgii_root().join("sync_connection_tokens.json")
+}
+
+/// Sync metrics log (rolling): `~/.orgii/sync-metrics.jsonl`.
+pub fn sync_metrics_log() -> PathBuf {
+    orgii_root().join("sync-metrics.jsonl")
+}
+
+/// Sync metrics log backup (rotated): `~/.orgii/sync-metrics.jsonl.1`.
+pub fn sync_metrics_log_backup() -> PathBuf {
+    orgii_root().join("sync-metrics.jsonl.1")
+}
+
+/// Global rules config: `~/.orgii/rules-config.json`.
+pub fn global_policies_config() -> PathBuf {
+    orgii_root().join("rules-config.json")
+}
+
+/// Workspace rules config: `{workspace}/.orgii/rules-config.json`.
+pub fn workspace_policies_config(workspace_path: &Path) -> PathBuf {
+    workspace_path.join(".orgii").join("rules-config.json")
+}
+
+/// MCP servers config (global): `~/.orgii/mcp-servers.json`.
+pub fn mcp_servers_config() -> PathBuf {
+    orgii_root().join("mcp-servers.json")
+}
+
+/// MCP servers config (workspace-level): `{workspace}/.orgii/mcp-servers.json`.
+pub fn workspace_mcp_servers_config(workspace_path: &Path) -> PathBuf {
+    workspace_path.join(".orgii").join("mcp-servers.json")
+}
+
+/// Global skills directory: `~/.orgii/skills/`.
+pub fn global_skills_dir() -> PathBuf {
+    orgii_root().join("skills")
+}
+
+/// File-based session registry root: `~/.orgii/sessions/`.
+///
+/// Crash-resilient per-session metadata files. Read/written by
+/// `agent_core::core::session::file_registry`.
+pub fn session_registry_dir() -> PathBuf {
+    orgii_root().join("sessions")
+}
+
+/// Per-session metadata file: `~/.orgii/sessions/<session_id>.json`.
+pub fn session_registry_file(session_id: &str) -> PathBuf {
+    session_registry_dir().join(format!("{}.json", session_id))
+}
+
+/// Chat image attachments: `~/.orgii/session-images/`.
+///
+/// Files are named by content hash for automatic deduplication.
+pub fn session_images_dir() -> PathBuf {
+    orgii_root().join("session-images")
+}
